Bulk C source fixes and analysis tooling: 867/886 functions now compile
Three automated source transformations across 650 C files:
- Constant pool resolver: 938 PTR_DAT/DAT references replaced with literals (225 files)
- Multiply-to-shift: 2445 `* power_of_2` converted to `<< N` with proper precedence (288 files)
- Pointer shift fix: 402 `(int)` casts added for pointer << N expressions (65 files)

Test harness: 43 PASS / 824 FAIL / 867 total (up from 39 PASS, 822 compilable)
Added 90+ analysis and debugging tools for failure triage.

+++ src/FUN_0601AB8C.c
@@ -19,7 +19,7 @@ void FUN_0601ab8c()
 
        *(int *)
 
-        (*(int *)(0x0605DE24 + ((*(int *)puVar1 << 1) + *(int *)0x0607EAE0) << 2) + 4);
+        (*(int *)(0x0605DE24 + ((*(int *)(int)puVar1 << 1) + *(int *)0x0607EAE0) << 2) + 4);
 
   return;
